+++ dashboard/src/components/Theme/Default.js
@@ -9,8 +9,18 @@ import colors from "../../colors";
 
 const DefaultTheme = createMuiTheme({
   palette: {
-    primary: colors.slateBlue,
-    secondary: colors.magenta,
+    primary: {
+      light: colors.paynesGrey[300],
+      main: colors.paynesGrey[500],
+      dark: colors.paynesGrey[600],
+      contrastText: "#F3F5F7",
+    },
+    secondary: {
+      light: colors.pistachio[300],
+      main: colors.pistachio[500],
+      dark: colors.pistachio[700],
+      contrastText: "#1D2237",
+    },
 
     // HACK: Reduce threshold white text while testing green theme
     getContrastText: color => {