+++ dashboard/src/components/Toolbar.js
@@ -1,51 +1,36 @@
 import React from "react";
-import compose from "lodash/fp/compose";
 import PropTypes from "prop-types";
-import { withRouter, routerShape } from "found";
 
 import AppBar from "material-ui/AppBar";
-import Avatar from "material-ui/Avatar";
 import MaterialToolbar from "material-ui/Toolbar";
 import Typography from "material-ui/Typography";
 import IconButton from "material-ui/IconButton";
-import Menu, { MenuItem } from "material-ui/Menu";
 import { withStyles } from "material-ui/styles";
-import MenuIcon from "material-ui-icons/Menu";
-import AppSearch from "./AppSearch";
-import { logout } from "../utils/authentication";
 
-import logo from "../assets/sensu-logo-white.png";
-import sampleAvatar from "../assets/sample-avatar.png";
+import MenuIcon from "material-ui-icons/Menu";
+import SearchIcon from "material-ui-icons/Search";
+
+import logo from "../assets/logo/wordmark/white.svg";
 
 const styles = theme => ({
   appBar: {
     transition: theme.transitions.create("width"),
   },
+  toolbar: {
+    marginLeft: -12, // Account for button padding to match style guide.
+    marginRight: -12,
+  },
   title: {
-    marginLeft: 24,
+    marginLeft: 20,
     flex: "0 1 auto",
   },
   grow: {
     flex: "1 1 auto",
   },
   logo: {
-    height: 15,
+    height: 16,
     marginRight: theme.spacing.unit * 1,
     verticalAlign: "baseline",
-  },
-  avatar: {
-    height: 32,
-    width: 32,
-    borderColor: "#fff",
-    borderWidth: 1,
-  },
-  search: {
-    marginRight: theme.spacing.unit,
-  },
-  hamburgerButton: {
-    // [theme.breakpoints.up("sm")]: {
-    //   display: "none",
-    // },
   },
 });
 
@@ -54,31 +39,6 @@ class Toolbar extends React.Component {
     // eslint-disable-next-line react/forbid-prop-types
     classes: PropTypes.object.isRequired,
     toggleToolbar: PropTypes.func.isRequired,
-    router: routerShape.isRequired,
-  };
-
-  state = {
-    menuAnchorEl: null,
-    menuOpen: false,
-  };
-
-  //
-  // Handlers
-
-  handleMenuButtonClick = event => {
-    this.setState({
-      menuOpen: !this.state.menuOpen,
-      menuAnchorEl: event.currentTarget,
-    });
-  };
-
-  handleMenuRequestClose = () => {
-    this.setState({ menuOpen: false });
-  };
-
-  handleLogout = async () => {
-    await logout();
-    this.props.router.push("/login");
   };
 
   //
@@ -89,13 +49,8 @@ class Toolbar extends React.Component {
 
     return (
       <AppBar className={classes.appBar}>
-        <MaterialToolbar>
-          <IconButton
-            className={classes.hamburgerButton}
-            onClick={toggleToolbar}
-            aria-label="Menu"
-            color="contrast"
-          >
+        <MaterialToolbar className={classes.toolbar}>
+          <IconButton onClick={toggleToolbar} aria-label="Menu" color="inherit">
             <MenuIcon />
           </IconButton>
           <Typography
@@ -105,31 +60,15 @@ class Toolbar extends React.Component {
             noWrap
           >
             <img alt="sensu logo" src={logo} className={classes.logo} />
-            Sensu
           </Typography>
           <div className={classes.grow} />
-          <AppSearch className={classes.search} />
-          <IconButton
-            aria-owns={this.state.menuOpen ? "profile-dropdown-menu" : null}
-            aria-haspopup="true"
-            onClick={this.handleMenuButtonClick}
-          >
-            <Avatar className={classes.avatar} src={sampleAvatar} />
+          <IconButton aria-label="Search" color="inherit">
+            <SearchIcon />
           </IconButton>
-          <Menu
-            id="profile-dropdown-menu"
-            anchorEl={this.state.menuAnchorEl}
-            open={this.state.menuOpen}
-            onRequestClose={this.handleMenuRequestClose}
-          >
-            <MenuItem>Profile</MenuItem>
-            <MenuItem>My account</MenuItem>
-            <MenuItem onClick={this.handleLogout}>Logout</MenuItem>
-          </Menu>
         </MaterialToolbar>
       </AppBar>
     );
   }
 }
 
-export default compose(withStyles(styles), withRouter)(Toolbar);
+export default withStyles(styles)(Toolbar);