+++ testing/e2e/check_scheduling_test.go
@@ -96,6 +96,7 @@ func TestCheckScheduling(t *testing.T) {
 	assert.NotEqual(t, count2, count3)
 
 	// Change the check schedule to cron
+	check.Interval = 0
 	check.Cron = "* * * * *"
 	err = sensuClient.UpdateCheck(check)
 	assert.NoError(t, err)